move piece-related types to own file
